fix: pageをSuspenseで囲むよう修正
